Rewrite the interface of `Target` for encapsulation.
Limiting direct access to the members of `Target` lead to a lot of
discoveries about the organisation of `Downloader` and `zck`
data and functions, which is why this commit is quite bigger than
initially expected.
Some code have been moved from `Downloader` to `Target` member functions
as they are mainly manipulating the state of `Target` without much
changes to `Downloader`, which basically is a sign that it's code
related to `Target`.